# src/main/resources/openapi/yanki-contract.yml
openapi: 3.0.3
info:
  title: Yanki Service API
  description: |
    API Monodero movil para transacciones
  version: 1.0.0
  contact:
    name: Equipo Bancario
    email: desarrollo@nttdata.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: https://localhost:8089/api/v1
    description: Servidor local de desarrollo

tags:
  - name: Yanki
    description: Operaciones con otros monederos Yanki

paths:
  /yanki:
    get:
      tags: [Yanki]
      summary: Listar todos los wallets Yanki
      description: Retorna una lista paginada de todos los wallets registrados
      operationId: findAll
      responses:
        '200':
          description: Lista de wallets obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YankiDto'

        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Yanki]
      summary: Crear nuevo wallet Yanki
      operationId: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiDto'
      responses:
        '201':
          description: Wallet de Yanki creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankiDto'

  /yanki/transfer:
    post:
      tags: [Yanki]
      summary: Realiza pago con wallet Yanki
      operationId: transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transactionDto'
      responses:
        '200':
          description: Pago realizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transactionDto'
        '404':
          $ref: '#/components/responses/NotFound'

  /yanki/link-debit-card:
    post:
      tags: [Yanki]
      summary: Asocia Yanki a una tarjeta de debito
      operationId: linkDebitCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiLinkRequest'
      responses:
        '200':
          description: Solicitud enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankiLinkRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    YankiDto:
      type: object
      properties:
        phone:
          type: string
          example: "976765433"
        documentNumber:
          type: string
          example: "87876523"
        imei:
          type: string
          example: "98976765654"
        email:
          type: string
          example: "yanki@email.com"
        debitCardNumber:
          type: string
          example: "1234-1234-1234-1234"
        balance:
          type: number
          example: 200.00

    transactionDto:
      type: object
      properties:
        mode:
          type: string
          enum: [YANKI, BANK]
        amount:
          type: number
          example: 200.00
        originProductId:
          type: string
          example: "0001-0004-0001-0001"
        destinyProductId:
          type: string
          example: "0001-0004-0001-0001"
        document:
          type: string
          example: "46491946"
        description:
          type: string
          example: "Pago por compra de un celular"
        tipo:
          type: string
          example: "PAYMENT"

    YankiLinkRequest:
      type: object
      properties:
        phone:
          type: string
          example: "976765433"
        debitCardNumber:
          type: string
          example: "87876523"


    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2023-01-15T10:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "El nro de tarjeta ya está registrado"
        path:
          type: string
          example: "/api/v1/debitcard"

  parameters:
    debitNumber:
      name: debitNumber
      in: path
      required: true
      description: ID único de tarjeta de débito
      schema:
        type: string
        format: string
        example: "9898-1234-1234-1234"

  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'